requirements.txt
python-dotenv
discord.py[voice]
youtube_dl @ git+https://github.com/ytdl-org/youtube-dl.git

